--- a/nntrainer/tensor/cl_operations/blas_kernel_strings.cpp
+++ b/nntrainer/tensor/cl_operations/blas_kernel_strings.cpp
@@ -677,84 +677,204 @@ const std::string &getDotClKernelFP16() {
 const std::string &getHgemmClNoTransKernel() {
   static const std::string hgemm_cl_noTrans_kernel_ =
     R"(
-        #pragma OPENCL EXTENSION cl_khr_fp16 : enable
-        __kernel void sgemm_cl_noTrans_fp16(const __global half* A, const __global half* B,
-                          __global half* C, unsigned int K, unsigned int lda, unsigned int ldb, unsigned int ldc) {
-            
-            unsigned int m = get_global_id(0);
-            unsigned int n = get_global_id(1);
-            float c = 0.0f;
-            for (unsigned int k = 0; k < K; ++k) {
-              float a, b;
-              a = A[m * lda + k];
-              b = B[k * ldb + n];
-              c += a * b;
-            }
-            C[m * ldc + n] = c;
-        })";
+
+    #pragma OPENCL EXTENSION cl_khr_fp16 : enable
+    #define TS 16
+    __kernel void sgemm_cl_noTrans_fp16(__global const half *A,
+                                        __global const half *B, __global half *C,
+                                        const int M, const int N, const int K) {
+      const int globalRow = get_global_id(1); // M dimension
+      const int globalCol = get_global_id(0); // N dimension
+
+      const int localRow = get_local_id(1);
+      const int localCol = get_local_id(0);
+      const int groupRow = TS * get_group_id(1);
+      const int groupCol = TS * get_group_id(0);
+
+      __local half Asub[TS][TS];
+      __local half Bsub[TS][TS];
+
+      float sum = 0.0f;
+
+      for (int t = 0; t < (K + TS - 1) / TS; ++t) {
+        const int tiledRowA = groupRow + localRow;
+        const int tiledColA = t * TS + localCol;
+
+        const int tiledRowB = t * TS + localRow;
+        const int tiledColB = groupCol + localCol;
+
+        // Load A
+        if (tiledRowA < M && tiledColA < K)
+          Asub[localRow][localCol] = A[tiledRowA * K + tiledColA];
+        else
+          Asub[localRow][localCol] = (half)0.0h;
+
+        // Load B
+        if (tiledRowB < K && tiledColB < N)
+          Bsub[localRow][localCol] = B[tiledRowB * N + tiledColB];
+        else
+          Bsub[localRow][localCol] = (half)0.0h;
+
+        barrier(CLK_LOCAL_MEM_FENCE);
+
+        for (int k = 0; k < TS; ++k)
+          sum += (float)(Asub[localRow][k]) * (float)(Bsub[k][localCol]);
+
+        barrier(CLK_LOCAL_MEM_FENCE);
+      }
+
+      if (globalRow < M && globalCol < N)
+        C[globalRow * N + globalCol] = (half)(sum);
+    }
+    )";
   return hgemm_cl_noTrans_kernel_;
 }
 
 const std::string &getHgemmClTransAKernel() {
   static const std::string hgemm_cl_transA_kernel_ =
     R"(
-        #pragma OPENCL EXTENSION cl_khr_fp16 : enable
-        __kernel void sgemm_cl_transA_fp16(const __global half* A, const __global half* B,
-                          __global half* C, unsigned int K, unsigned int lda, unsigned int ldb, unsigned int ldc) {
-            
-            unsigned int m = get_global_id(0);
-            unsigned int n = get_global_id(1);
-            float c = 0.0f;
-            for (unsigned int k = 0; k < K; ++k) {
-              float a, b;
-              a = A[k * lda + m];
-              b = B[k * ldb + n];
-              c += a * b;
-            }
-            C[m * ldc + n] = c;
-        })";
+      for (int t = 0; t < (K + TS - 1) / TS; ++t) {
+        const int tiledRowA = t * TS + localCol;
+        const int tiledColA = groupRow + localRow;
+
+        const int tiledRowB = t * TS + localRow;
+        const int tiledColB = groupCol + localCol;
+
+        // Load Aᵗ (A[col * M + row])
+        if (tiledRowA < K && tiledColA < M)
+          Asub[localRow][localCol] = A[tiledRowA * M + tiledColA];
+        else
+          Asub[localRow][localCol] = (half)0.0h;
+
+        // Load B (K x N)
+        if (tiledRowB < K && tiledColB < N)
+          Bsub[localRow][localCol] = B[tiledRowB * N + tiledColB];
+        else
+          Bsub[localRow][localCol] = (half)0.0h;
+
+        barrier(CLK_LOCAL_MEM_FENCE);
+
+        for (int k = 0; k < TS; ++k)
+          sum += (float)(Asub[localRow][k]) * (float)(Bsub[k][localCol]);
+
+        barrier(CLK_LOCAL_MEM_FENCE);
+      }
+
+      if (globalRow < M && globalCol < N)
+        C[globalRow * N + globalCol] = (half)(sum);
+    }
+    )";
   return hgemm_cl_transA_kernel_;
 }
 
 const std::string &getHgemmClTransBKernel() {
   static const std::string hgemm_cl_transB_kernel_ =
     R"(
-        #pragma OPENCL EXTENSION cl_khr_fp16 : enable
-        __kernel void sgemm_cl_transB_fp16(const __global half* A, const __global half* B,
-                          __global half* C, unsigned int K, unsigned int lda, unsigned int ldb, unsigned int ldc) {
-            
-            unsigned int m = get_global_id(0);
-            unsigned int n = get_global_id(1);
-            float c = 0.0f;
-            for (unsigned int k = 0; k < K; ++k) {
-              float a, b;
-              a = A[m * lda + k];
-              b = B[n * ldb + k];
-              c += a * b;
-            }
-            C[m * ldc + n] = c;
-        })";
+    #pragma OPENCL EXTENSION cl_khr_fp16 : enable
+    #define TS 16
+    __kernel void sgemm_cl_transB_fp16(__global const half *A,
+                                      __global const half *B, __global half *C,
+                                      const int M, const int N, const int K) {
+      const int globalRow = get_global_id(1); // M
+      const int globalCol = get_global_id(0); // N
+
+      const int localRow = get_local_id(1);
+      const int localCol = get_local_id(0);
+      const int groupRow = TS * get_group_id(1);
+      const int groupCol = TS * get_group_id(0);
+
+      __local half Asub[TS][TS];
+      __local half Bsub[TS][TS];
+
+      float sum = 0.0f;
+
+      for (int t = 0; t < (K + TS - 1) / TS; ++t) {
+        const int tiledRowA = groupRow + localRow;
+        const int tiledColA = t * TS + localCol;
+
+        const int tiledRowB = groupCol + localCol;
+        const int tiledColB = t * TS + localRow;
+
+        // Load A (M x K)
+        if (tiledRowA < M && tiledColA < K)
+          Asub[localRow][localCol] = A[tiledRowA * K + tiledColA];
+        else
+          Asub[localRow][localCol] = (half)0.0h;
+
+        // Load Bᵗ (B[col * K + row])
+        if (tiledRowB < N && tiledColB < K)
+          Bsub[localRow][localCol] = B[tiledRowB * K + tiledColB];
+        else
+          Bsub[localRow][localCol] = (half)0.0h;
+
+        barrier(CLK_LOCAL_MEM_FENCE);
+
+        for (int k = 0; k < TS; ++k)
+          sum += (float)(Asub[localRow][k]) * (float)(Bsub[k][localCol]);
+
+        barrier(CLK_LOCAL_MEM_FENCE);
+      }
+
+      if (globalRow < M && globalCol < N)
+        C[globalRow * N + globalCol] = (half)(sum);
+    }
+    )";
   return hgemm_cl_transB_kernel_;
 }
 
 const std::string &getHgemmClTransABKernel() {
   static const std::string hgemm_cl_transAB_kernel_ =
     R"(
-        #pragma OPENCL EXTENSION cl_khr_fp16 : enable
-        __kernel void sgemm_cl_transAB_fp16(const __global half* A, const __global half* B,
-                          __global half* C, unsigned int K, unsigned int lda, unsigned int ldb, unsigned int ldc) {
-            
-            unsigned int m = get_global_id(0);
-            unsigned int n = get_global_id(1);
-            float c = 0.0f;
-            for (unsigned int k = 0; k < K; ++k) {
-              float a, b;
-              a = A[k * lda + m];
-              b = B[n * ldb + k];
-              c += a * b;
-            }
-            C[m * ldc + n] = c;
-        })";
+    #pragma OPENCL EXTENSION cl_khr_fp16 : enable
+    #define TS 16
+    __kernel void sgemm_cl_transAB_fp16(__global const half *A,
+                                        __global const half *B, __global half *C,
+                                        const int M, const int N, const int K) {
+      const int globalRow = get_global_id(1); // M
+      const int globalCol = get_global_id(0); // N
+
+      const int localRow = get_local_id(1);
+      const int localCol = get_local_id(0);
+      const int groupRow = TS * get_group_id(1);
+      const int groupCol = TS * get_group_id(0);
+
+      __local half Asub[TS][TS];
+      __local half Bsub[TS][TS];
+
+      float sum = 0.0f;
+
+      for (int t = 0; t < (K + TS - 1) / TS; ++t) {
+        const int tiledRowA = t * TS + localCol;
+        const int tiledColA = groupRow + localRow;
+
+        const int tiledRowB = groupCol + localCol;
+        const int tiledColB = t * TS + localRow;
+
+        // Load Aᵗ (K x M)
+        if (tiledRowA < K && tiledColA < M)
+          Asub[localRow][localCol] = A[tiledRowA * M + tiledColA];
+        else
+          Asub[localRow][localCol] = (half)0.0h;
+
+        // Load Bᵗ (N x K)
+        if (tiledRowB < N && tiledColB < K)
+          Bsub[localRow][localCol] = B[tiledRowB * K + tiledColB];
+        else
+          Bsub[localRow][localCol] = (half)0.0h;
+
+        barrier(CLK_LOCAL_MEM_FENCE);
+
+        for (int k = 0; k < TS; ++k)
+          sum += (float)(Asub[localRow][k]) * (float)(Bsub[k][localCol]);
+
+        barrier(CLK_LOCAL_MEM_FENCE);
+      }
+
+      if (globalRow < M && globalCol < N)
+        C[globalRow * N + globalCol] = (half)(sum);
+    }
+
+    )";
   return hgemm_cl_transAB_kernel_;
 }