This pull request adds a highly optimized half-precision General Matrix Multiplication (GEMM) kernel developed for OpenCL. 

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

## Result

#### PC (x86)
|engine|avg|
|------|---|
|CPU|116ms|
|GPU (prev)|3845 ms|
|GPU|63.05 ms|

#### Android (aarch64)
|engine|avg|
|------|---|
|CPU|214.2ms|
|GPU (prev)|26572.3ms|
|GPU|503.35 ms|

## Note

This is a profiling result with the following matrix size.
- M: 1024
- K: 3072
- N: 3072